Move google docs list filter to list.js file.
